chore: refactoring + tests
