+++ packages/blockchain-sdk-polygon/src/auth/polygon-ed25519-authentication-facade-factory.ts
