+++ packages/blockchain-sdk-polygon/src/auth/polygon-ed25519-token-validator.ts
