+++ packages/blockchain-sdk-polygon/src/index.ts
@@ -1,0 +1,6 @@
+export * from './auth/polygon-ed25519-authentication-facade-factory';
+export * from './wallet-adapter/dialect-polygon-wallet-adapter-wrapper';
+export * from './wallet-adapter/node-dialect-polygon-wallet-adapter';
+export * from './wallet-adapter/dialect-polygon-wallet-adapter.interface';
+export * from './sdk/sdk';
+export * from './sdk/constants';
