+++ packages/blockchain-sdk-polygon/src/wallet-adapter/dialect-polygon-wallet-adapter.interface.ts
